Sound and HV Frequency implemented simultaneously
HV Frequency not using interrrupts
Sound ain't using interrupts either. However, TA1CCR2 interrupt has been
implemented to count how many "sound cycles" run before turning sound
off.

--------------- SOUND NOT WORKING YET --------------------
Should debug TA1CCR2 Interrupts

+++ Firmware/FrequencySetup/lcd.c
@@ -57,8 +57,8 @@ void lcd_print(char *s){
     }
 }
 
-void lcd_print_number(unsigned int number){
-	unsigned int n = number;
+void lcd_print_number(unsigned long number){
+	unsigned long n = number;
 	unsigned int nOut[16]; //At most 16 digits per number
 	unsigned int cnt = 0;
 	int i;
