+++ Firmware/FrequencySetup/lcd.h
@@ -17,7 +17,7 @@ void data_read(void);
 void send_command(unsigned char cmd);
 void send_data(unsigned char data);
 void lcd_print(char *s);
-void lcd_print_number(unsigned int number);
+void lcd_print_number(unsigned long number);
 void lcd_init(void);
 void lcd_clear(void);
 void lcd_gotoRow(unsigned int row);
